Fixed warning
git-svn-id: http://miranda.googlecode.com/svn/trunk@9471 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/AimOscar/avatars.cpp
@@ -180,7 +180,7 @@ void  CAimProto::get_avatar_filename(HANDLE hContact, char* pszDest, size_t cbLe
 	init_custom_folders();
 
 	char* path = (char*)alloca( cbLen );
-	if (hAvatarsFolder == NULL || FoldersGetCustomPath(hAvatarsFolder, path, cbLen, ""))
+	if (hAvatarsFolder == NULL || FoldersGetCustomPath(hAvatarsFolder, path, (int)cbLen, ""))
 	{
         char *tmpPath = Utils_ReplaceVars("%miranda_avatarcache%");
 		tPathLen = mir_snprintf(pszDest, cbLen, "%s\\%s", tmpPath, m_szModuleName);
